Import wayland-1.19.0
Comment:
Display server protocol

Description:
Wayland is intended as a simpler replacement for X, easier to develop and
maintain.

Wayland is a protocol for a compositor to talk to its clients as well as
a C library implementation of that protocol. The compositor can be a
standalone display server running on Linux kernel modesetting and evdev
input devices, an X application, or a wayland client itself. The clients can
be traditional applications, X servers (rootless or fullscreen) or other
display servers.

Port notes:
This work is only possible because to FeeBSD's wayland patches and sthen@'s
work on devel/epoll-shim. From my position it is currently needed as a
dependency.

It should give other developers the possibility to play with wayland e.g. to
build Xenocara with XWayland support.

New category "wayland" to relieve x11/devel.

Feedback and OK fcambus@ daniel@