drop maintainer on sysutils/dtb
Updating this port has a high chance of breaking drivers as breaking
changes are frequently made to device trees in linux.  For example the
description of BeagleBone Black hardware changed in linux 5.15 to be
compatible with a new cpsw driver in linux.  This would break cpsw(4) if
the port were updated to 5.15.