update to libexttextcat-3.4.6 which has some more classifiers
- set USE_GMAKE so the upstream install target works properly, which
means we don't need to handroll installing the lm files in post-install,
which means they get installed to the correct directory. tracked
down from a problem with Dovecot reported by Johan Huldtgren.

- build/install the testtextcat program that's used in regression tests,
as a simple way to see what the library is doing (run it with a path
to fpdb.conf, feed on stdin, it tries to identify the language.
looks like you need to be in share/libexttextcat for it to pick up
the lm files correctly)