Adjust perl.port.mk so options will be picked up in more places
Export some ExtUtils::MakeMaker and Module::Build recognized environment
variables so they are picked up even if we don't run the configure
script directly.

This is useful for things that are not perl modules themselves, but include
perl modules, some of which seem to configure them at build time instead
of configure.

This is the first step to moving perl to have a vendor lib that
ports install into that is separate from where CPAN will install things.

Thanks to kmos@ for a bulk build that didn't turn up anything untoward