Update to NRPE 4.0.2, based on a diff from trondd
This has slightly less ancient TLS support (allowing us to get rid of
use of openssl to build) but is not compatible with older versions.

The warning remains; if you are still using this it is recommended that
you don't:

WARNING: NRPE is deprecated upstream and no longer receives bug fixes or
new features. For some of the checks on remote servers that you might
otherwise make using NRPE, you may like to consider "manubolon-snmp" and
an SNMP daemon instead.