switch lablgtk3 from to the .tbz release which includes additional metadata
The .tbz archive contains metadata which allows "ocamlfind list" to
correctly show the version rather than showing "n/a".

This issue was found by Yoza Toda back in March while attempting to update
math/coq to 8.13.1. That version of coq apparently uses "ocamlfind list" to
check for the presence of lablgtk3.

Other dune ports may have the same issue, although an alternative fix
could also be to get dune itself updated to 2.8.0 or later as that version
of dune includes a different fix for this issue.