Backout previous opendnssec changes to cope with it wanting to use constants from ldns for a different purpose, that ldns doesn't really want to expose on non-EdXXX builds.
Bodge it in ldns instead by supporting the same USE_EDxxx cpp macros
that opendnssec uses. Looks to me like ldns are trying to replace them
with something harder to mess about with.

Hopefully the various parts of nlnetlabs responsible for these two
pieces of software will talk to each other and figure something out.

Hopefully this should fix ed25519 with opendnssec as found by an
anonymous reporter on ports@.