Import audio/rnnoise 0.0.0.20210312
It is a dependency of an upcoming port, but it will also be picked up by
audio/mumble to enable optional noise suppression there as well.

Feedback sdk tb phessler
OK sdk
---
RNNoise is a noise suppression library based on a recurrent neural network.
A description of the algorithm is provided in the following paper:

J.-M. Valin, A Hybrid DSP/Deep Learning Approach to Real-Time Full-Band Speech
Enhancement, Proceedings of IEEE Multimedia Signal Processing (MMSP) Workshop,
arXiv:1709.08243, 2018.
https://arxiv.org/pdf/1709.08243

While it is meant to be used as a library, the code for a simple command-line
tool is provided as an example. It operates on RAW 16-bit (machine endian) mono
PCM files sampled at 48 kHz.