Import devel/range-v3 0.11.0
Feedback OK rsadowksi
---
Range library for C++14/17/20.  This code was the basis of
[a formal proposal](https://ericniebler.github.io/std/wg21/D4128.html) to
add range support to the C++ standard library.  That proposal evolved
through a Technical Specification, and finally into
[P0896R4 "The One Ranges Proposal"](https://wg21.link/p0896r4) which was
merged into the C++20 working drafts in November 2018.

Ranges are an extension of the Standard Template Library that makes its
iterators and algorithms more powerful by making them _composable_.
Unlike other range-like solutions which seek to do away with iterators,
in range-v3 ranges are an abstration layer _on top_ of iterators.