update to SDL 2.0.18 similar diff proposed by Brad
I added a fix for audio assertion failure and removed the code where SDL
would pretend to be Linux for compatibility reasons with FNA games - this
has been addressed upstream and the need for this workaround is gone.

Many consumers tested with build and runtime; no issues found.

Release notes at:
https://discourse.libsdl.org/t/sdl-2-0-18-released/33768