Update ssh-ldap-helper to 8.0p1
This is the latest version of this tool, based on what Rocky Linux ships.
Red Hat have deprecated this, but their proposed alternative, SSSD is not
currently viable.

The update is ncessary since the previous version (which wasn't updated
since import in 2012) is based on an ancient SSH version that isn't ready
for opaque libcrypto structs.

Untested, but this has a better chance at working than a package that
doesn't build.

ok landry