Remove a local patch from mutter which changes a memcpy into memmove.
I can no longer trigger the segfault which prompted aja to add this
patch to our port. It is unclear what caused the original issue.
We should not carry a local patch which differs from upstream code
with no observable benefit. If the problem resurfaces we will be
able to debug it further.

ok ajacoutot@