update to zsh-5.8.1, from Matthew Martin
This is a stable security release with a few bug fixes, including one
for CVE-2021-45444, a vulnerability in prompt expansion which could be
exploited through e.g. VCS_Info to execute arbitrary shell commands
without a user's knowledge. All sites are encouraged to update from
zsh 5.8. A partial work-around which can be applied within a running
shell is provided in the source distribution for those who are unable
to update their shell binaries.

https://www.zsh.org/mla/announce/msg00133.html