build yubikey-manager with its own copy of the current 0.8 version of py3-fido.
should be a noop. if it doesn't work currently or only partially works
(like mine which times out usb transactions for many operations) then
this won't help, if it does work then this shouldn't break things.
but this gets ykman out of the way of updates to the main py-fido port.