Fix ENODATA usage on non-clang arches
ENODATA from /usr/include/c++/v1/errno.h is only defined with base-clang
(I didn't follow the header maze) and a few other ports already replace
ENOADATA with ENOENT or ENOATTR, so do the same except just pass it along as
flags to avoid further patching.

ENOENT is used in fuse-zip, ENOATTR is not, so pick the latter.

OK kmos