Use ports-gcc to fix sparc64 build, typofix python dependency
> /tmp/print-dict-f43ef2.s:4649: Error: operation combines symbols in different segments
> clang-13: error: assembler command failed with exit code 1 (use -v to see invocation)

Apply the usual COMPILER fix.

update-plist still puts a python dir into PLIST-main which is wrong but I
don't know how to convince it to account for it in PLIST-python... by not
listing it at all since it belongs to the python package which is now a
proper dependency.

OK kmos