Import twmn 1.2
Twmn is a dmenu style desktop notification system.

Notifications are shown in a one-line bar called the notification slide.
They can be navigated through and activated with shortcuts.

* twmnc: is a command line tool to send notifications to twmnd.
* twmnd: is a daemon listening to notification requests and showing them
one after another. It is also compatible with "notify-send".

Tested on amd64. An example config can be found in the README.md in the
examples dir.

With help from op@ bcallah@

OK op@ bcallah@