Update to unison-2.52.0
This release introduces a new wire protocol, a new archive format, and
feature negotiation. In particular, once two machines have been upgraded
to unison-2.52.0 or later, it is possible to use unison compiled with
different ocaml versions. Summary of changes can be found at
https://github.com/bcpierce00/unison/blob/v2.52.0/NEWS.md.

This release is compatible with:
- unison 2.52 or newer. You do not have to pay any attention to OCaml
  compiler versions
- unison 2.51 if both versions are compiled with same OCaml compiler
  version
- unison 2.48 if both versions are compiled with same OCaml compiler
  version

Comments and OK sthen@