import doulos-cipher-1.020
The Cipher music notation is used throughout Indonesia and China
for all kinds of music. The font is not intended for general
orthographic use, although it is usable for song lyrics in simple
Roman writing systems. This solution is primarily intended for
fairly straightforward songbooks for singers, not for complex
instrumental music. It can handle harmony lines fine, but not a lot
of intricacies or subtleties (e.g. no staccato).

from George Rosamond (MAINTAINER) with tweaks from Yifei Zhan, thank
you!  OK sthen