update luarocks to 3.9.0
No changelog available this time (yet?), but looking at the commits
between the last release it's a small update with only a few things that
affects us.  Tested with a WIP luarocks.port.mk and as a user
(luarocks-5.1 install --local XYZ) with both "pure lua" rocks and
libraries that needs C bits.

while here add a comment explaining why NO_TESTS so I don't forget that
it needs busted.