Import clidle v0.0.0-20220208035357
This is the popular wordle game for the terminal, written in go.

Guess the word in six tries. Each guess must be a valid five-letter
word. Hit enter to submit. After each guess, the color of the tiles will
change to show how close your guess was to the word.

With suggestions from op@

ok op@