split Godot into two packages: engine alone (-main) and editor (-tools)
A Godot build with the tools enabled is not really that great to run
standalone games due to the various extra checks that the engine does,
so disable the tools (i.e. the editor) in the -main package, but provide
a -tools package for folks that wants to develop games.

While here use target=release instead of debug_release for the -main
package: comments in the SConstruct file hints that debug_release
enables check useful only for development.

discussed many times with various people, tested by few (thanks!)
ok/improvements from sthen@