Python 3.10: unbreak _hashlib build
The build was broken in the update to 3.10.2 due to the upstream commit
https://github.com/python/cpython/commit/3ce6945f5f434806eea700eb5ff1bed6d39395e1
This seems to have escaped notice of the maintainers because the build
system only whines and carries on instead of failing (although this was
also visible in the PLIST). Avoid use of undefined short names and NIDs
and do not use an inexistent flag.

ok sthen