try to unbreak games/godot,-main on gcc arches
platform/x11/detect.py tries to create export templates when building
with GNU ld and tools=no.  GNU ld in base is too old for that and it's
not straightforward to enable them using llvm (upstream says it doesn't
work) so just disable them; they're not that useful for us anyway.

bumping is just a pro-forma in this case: on llvm arches there are no
changes and on gcc-arches the previous version didn't built.

(while here reformat the comment at the top of patch-platform_x11_detect_py)