Add a QML just-in-time compiler comment
"Previously a port using this library could decide whether or not to allow the
JIT to work by setting or not setting USE_WXNEEDED. The patch means that this
is no longer possible but it means that someone setting the kern.wxabort debug
sysctl can still run that software. (It's not a security sysctl because the w+x
mapping is rejected whichever way it is set)." -- sthen@

Comment from sthen@, thanks!