lang/sbcl: disable contrib tests
tests for the contrib packages are run during the build.  sb-concurrency
has failed in bulks three times in a row (reported by naddy) but I can't
reproduce the failure locally so disable the tests.  This is a quick
hack to allow sbcl,threads to build, hence not waiting for the
maintainer.  (I'm working on a better approach that runs the contrib
tests during `make test'.)