allow the ssh master to start interactively, so that it's even possible to run without ssh keys or with a passphrase.
allow up to 10 seconds for startup so that the prompts do not
overlap each other (after that, the host will be dead as usual).

paraphrased from naddy@ "it's unusual to allow root ssh with a password,
but that means that problems with the configuration will show up straight
away" which is a good thing.