Enable the "rock" disk storage backend for www/squid, and mention in README that to use it you will need to raise net.unix.dgram.sendspace above the default of 2048 (it's trying to sendmsg a little above 4k). Add @sample for the default dirs used for ipc communications.
Only lightly tested, but it's easier to test further if it's in the
default build.

Thanks tb@ for assistance in tracking down the limit that I was hitting.

For the benefit of future searches, the log message from squid when hitting
this case is:

src/comm/Write.cc(120) HandleWrite: FD 30 write failure: (40) Message too long