mail/aerc: update to 0.10.0 for real
last update only bumped "V" and not really updated the port.  Appeding
"pl0" to the package version to have a differently-named DISTNAME and
avoid breakage.

Thanks to James Cook (falsifian [at] falsifian org) for noticing the
issue, semarie and sthen for confirming that without a name change it
would break in bulks.  OK sthen@