import fonts from George Rosamond. ok op@
---
Jaldi is a libre Devanagari typeface family designed as a complement
to the Asap font. Jaldi is a contemporary sans-serif non-modular
Devanagari family with subtle rounded corners designed by Nicolas
Silva and Pablo Cosgaya.
---
Aref Ruqaa is an Arabic typeface that aspires to capture the essence
of the classical Ruqaa calligraphic style.

The Arabic part of Aref Ruqaa is designed by the Arabic calligrapher
and type designer Abdullah Aref, while the Latin part is based on
AMS Euler, but spaced for regular text rather than mathematics.
---
Vazirmatn is a Persian/Arabic font project that started in 2015
under the name of Vazir with the idea of a new simple and legible
typeface suitable for web pages and applications.
---