import port/devel/riscv32-esp-elf, ok sthen@
The riscv32-esp-elf port is an ESP32 GNU cross compiler suite, configured for
the riscv32-esp-elf target.

ESP32-C3 is a single-core Wi-Fi and Bluetooth 5 (LE) microcontroller SoC, based
on the open-source RISC-V architecture. It strikes the right balance of power,
I/O capabilities and security, thus offering the optimal cost-effective solution
for connected devices. The availability of Wi-Fi and Bluetooth 5 (LE)
connectivity not only makes the device's configuration easy, but it also
facilitates a variety of use-cases based on dual connectivity.