Update prometheus to v2.36.2
Switch the port to use the official prometheus-web-ui package to avoid
the nightmare of building the react project on the ports infrastructure.

Include a patch from an open pull request that works around the issue
of TSDB on systems with no unified buffer cache by using mmap both for
read and writes. For more info about patch-mmap_openbsd check out:
   https://github.com/prometheus/prometheus/pull/9085
and
   https://github.com/prometheus/prometheus/issues/8877
OK sthen@