Backport a fix for some integer multiplications on octeon
Integer multiplications where both integers are less than 64-bit but the
multiplication was greater than 64-bit were handled using an int128_t
type on octeon, with inconsistent results.  The simplest fix in this
case is to not use an int128_t type, and fallback to the default codepath.
I pushed a fix for this upstream, this backports that fix.

Problem discovered by the openssl-ruby tests
Initial debugging by tb@ and claudio@
Access to octeon thanks to bluhm@
Further debugging and fix by me
Testing by tb@ and visa@

OK tb@