import guile3-3.0.8
This is version 3.0 of Guile, Project GNU's extension language library.
Guile is an implementation of the Scheme programming language, packaged
as a library that can be linked into applications to give them their own
extension language.  Guile supports other languages as well, giving
users of Guile-based applications a choice of languages.

gkoehler@ tested it on powerpc64 and confirmed that it's broken on
powerpc.

ok sthen@