update to 1.31.0-UNSTABLE
Notable changes since 1.29.0:
- MD: Fixed glitches and hang in the Europe release of "Donald in Maui Mallard".
- PC-FX: Expanded external backup memory from 32KiB to 128KiB, to match the
amount of memory in an FX-BMP.
- PC-FX: External backup memory will now be saved to/loaded from its own file,
instead of being appended to the internal backup memory in the same file. For
backwards compatibility, there is support for initially loading data from a
combined 32KiB+32KiB file; however, the additional 96KiB of external backup
memory will be unusable by the game in this case due to existing filesystem
structures.
- SASPlay: Added support for "Air Walkers".
- SNES-Faust: Added mouse emulation.
- SNES-Faust: Fixed color math glitch in "Final Fantasy V" when Exdeath is
holding the opaque white sphere party hostage.
- SNES-Faust: Fixed broken ExLoROM support, used by a "Heartthrob Memorial"
translation patch.
- SS: Added experimental support for ST-V games that don't require decryption
chips or special hardware. Refer to the ST-V section of the documentation for
more details and a list of explicitly supported games.
- SS: Fixed missing and glitchy graphics in "Houkago Ren'ai Club: Koi no Etude".
- SS: Fixed graphical glitches in "Radiant Silvergun" at the start of a new game
begun after interrupting the attract mode, or game over, while a fancy rotating
background was present, such as during the AKA-O battle.