Python tidying:
in one of the 3.9.x updates we accidentally lost the "Make sure
CPPFLAGS/LDFLAGS passed to ./configure don't end up -in
PY_CPPFLAGS/PY_LDFLAGS and thus leaking to python extensions" in
a merge.

The original problem this was fixing was a path to the ports WRKDIR
appearing in the built package. (which showed up in "python3.9 -m
sysconfig|grep LDSH" etc).

However that problem no longer shows up in 3.9, so it seems sensible
to sync 3.10 to what we have now for 3.9 and remove the patch, and
update CHANGES.OpenBSD (removing the bogus entry for 3.9, and the
no-longer-needed one for 3.10).

ok kmos@