update FNA suite to 22.08
Link mojoshader statically in fna3d and lose mojoshader dependendencies
in the process. This is the way upstream recommends it and uses it on
other platforms.

FAudio upstream has gotten rid of the built-in GSTREAMER use.

Tested with a large library of FNA and XNA games. Fallout from the
changes is minimal and doesn't affect any of the major or more popular
games.

On the positive side, the crashes that I reported in [1] are gone with
this update.

Changenotes in [2], [3], and [4].

[1] https://marc.info/?l=openbsd-bugs&m=165136879315845&w=2
[2] https://github.com/FNA-XNA/FNA/releases
[3] https://github.com/FNA-XNA/FAudio/releases
[4] https://github.com/FNA-XNA/FNA3D/releases