update textproc/libgrapheme to 2.0.1
No API/ABI changes to the library itself just some hardening, tweaks to
the build helpers and to the build system.  Now it needs less massagging
to accomodate for our requirements on ports' libraries.

Building like upstream does with -nostdlib seems to break `update-plist'
so I'm avoiding it for the moment; not a big deal anyway.

ok jan@