Import print/weasyprint 56.1.
WeasyPrint is a smart solution helping web developers to create PDF
documents. It turns simple HTML pages into gorgeous statistical
reports, invoices, tickets..

From a technical point of view, WeasyPrint is a visual rendering engine
for HTML and CSS that can export to PDF. It aims to support web
standards for printing.

It is based on various libraries but not on a full rendering engine
like WebKit or Gecko. The CSS layout engine is written in Python,
designed for pagination, and meant to be easy to hack on.

www: https://weasyprint.org/

ok sthen@ semarie@