Tor Browser: update to 12.0a3
Yes, that's an alpha release, which is not ideal. Unfortunately however,
Tor Browser's last stable release is based on an unsupported version of
Firefox ESR which does not build with the newest version of Rust (which
we already have).

If you don't want to use an alpha release of Tor Browser, it's probably
best to run OpenBSD 7.2 for another month until Tor Browser 12.0
is released.

Encouraged by landry@ tb@ george@
OK landry@