import nushell-0.70.0
The goal of nushell is to take the Unix philosophy of shells,
where pipes connect simple commands together, and bring it
to the modern style of development.  Thus, rather than being
either a shell, or a programming language, nushell connects
both by bringing a rich programming language and a full-featured
shell together into one package.

ok tb@
with help from semarie@ and tb@