Patch Modules/readline.c to workaround an issue with the readline module with Python 3.9+'s definition of PyMODINIT_FUNC and our readline headers https://github.com/openbsd/src/blob/master/gnu/lib/libreadline/rlstdc.h#L41
(Due to some other strangeness in the build system, on 3.9 it rebuilt a
working version after the first attempt failed and the file was moved out
of the way, but that no longer happens with 3.10+ where readline just
failed).

Fix up PLISTs and WANTLIB for Python 3.11 while there.

Joint work with landry@ tb@ kmos@, ok tb kmos