add back a version of Postfix from the 3.5 "past stable" branch, for two reasons:
- despite static linking, there are still some problems with the flavours
that pull in a dependency that uses libressl, at least when that dep is
making a TLS connection (e.g. to database servers on another machine).

- to provide a version of Postfix that is still able to use libressl
rather than pulling in openssl, as upstream's more recent stable versions
require. as a post on mailop points out in relation to a recent issue
with OpenSSL, "MTAs usually do a lot of outbound TLS acting as clients
to remote servers, but opportunistically (disabled cert validation)"
so SMTP clients will see many certs that might not make it through
CA validation checks.

ok with Brad (maintainer)