import mail/stalwart 0.2.0
bundles 3 rust ports for:

* Stalwart JMAP is an open-source JSON Meta Application Protocol server
designed to be secure, fast, robust and scalable. JMAP is a modern
protocol for synchronising data such as mail, calendars, or contacts
that makes much more efficient use of network resources.

* Stalwart IMAP is an open-source Internet Message Access Protocol server
designed to be secure, fast, robust and scalable. A JSON Meta
Application Protocol (JMAP) backend such as Stalwart JMAP is required to
use Stalwart IMAP (in other words, Stalwart IMAP is an imap4-to-jmap
proxy).

* CLI to administer a Stalwart JMAP server. Allows to:
- manage domains (including DKIM)
- manage accounts
- manage lists (lightweight mailing lists)
- manage groups (lightweight aliases)
- import accounts & mailboxes

cf https://jmap.io for more details on JMAP protocol/specifications. A
bit rough around the edges, but works in simple testing with
mutt/thunderbird against the imap proxy, and should soon work with meli
against the jmap server. A README provides basic info on how to run the
whole thing.

thanks to semarie@ for help fighting with cargo modules.

ok sthen@